refactor: remove old native interceptor and renamed new-native to take place
