+++ packages/hoppscotch-selfhost-desktop/src/main.ts
@@ -14,7 +14,6 @@ import { appWindow } from "@tauri-apps/api/window"
 import { stdFooterItems } from "@hoppscotch/common/platform/std/ui/footerItem"
 import { stdSupportOptionItems } from "@hoppscotch/common/platform/std/ui/supportOptionsItem"
 import { ioDef } from "./platform/io"
-import { NewNativeInterceptorService } from "@platform/interceptors/new-native"
 
 const headerPaddingLeft = ref("0px")
 const headerPaddingTop = ref("0px")
@@ -50,7 +49,6 @@ const headerPaddingTop = ref("0px")
       default: "native",
       interceptors: [
         { type: "service", service: NativeInterceptorService },
-        { type: "service", service: NewNativeInterceptorService },
         { type: "standalone", interceptor: proxyInterceptor },
       ],
     },
