+++ packages/hoppscotch-selfhost-desktop/src/platform/interceptors/native/persisted-data.ts
